% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPfit-package.R
\name{scale_norm}
\alias{scale_norm}
\title{Scale variable into normal range 0, 1}
\usage{
scale_norm(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector
}
\description{
Perform calculation:
(x - min(x)) / (max(x) - min(x))
}
\examples{
scale_norm(x = c(-1, 4, 10, 182))
}
