% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_matrix.R
\name{corr_matrix}
\alias{corr_matrix}
\title{Power Exponential or Matern Correlation Matrix}
\usage{
corr_matrix(X, beta, corr = list(type = "exponential", power = 1.95))
}
\arguments{
\item{X}{the (\code{n x d}) design matrix}

\item{beta}{a (\code{d x 1}) vector of correlation hyper-parameters in
\eqn{(-\infty, \infty)}}

\item{corr}{a list that specifies the \code{type} of correlation function
along with the smoothness parameter. The default corresponds to power
exponential correlation with smoothness parameter "\code{power=1.95}". One
can specify a different power (between 1.0 and 2.0) for the power
exponential, or use the Matern correlation function, specified as
\code{corr=list(type = "matern", nu=(2*k+1)/2)}, where \eqn{k \in
\{0,1,2,...\}}}
}
\value{
The (\code{n x n}) correlation matrix, R, for the design matrix
(\code{X}) and the hyper-parameters (\code{beta}).
}
\description{
Computes the power exponential or Matern correlation matrix for a set of
\emph{n} design points in \emph{d}-dimensional input region and a vector of
\emph{d} correlation hyper-parameters (beta).
}
\details{
The power exponential correlation function is given by

\eqn{R_{ij} = \prod_{k=1}^{d} \exp({-10^{\beta_k}|x_{ik}-x_{jk}|^{power}})}{
R_{ij} = \prod exp(-10^\beta*|x_{i}-x_{j}|^{power})}. \cr

The Matern correlation function given by Santner, Williams, and Notz (2003)
is

\eqn{R_{ij} = \prod_{k=1}^{d}
\frac{1}{\Gamma(\nu)2^{\nu-1}}(2\sqrt{\nu}|x_{ik} - }{R_{ij} = \prod
\frac{1}{\Gamma(\nu)2^{\nu-1}}(2\sqrt{\nu}|x_{ik} - x_{jk}|10^{\beta_k})^\nu
\kappa_{\nu}(2\sqrt{\nu}|x_{ik} -
x_{jk}|10^{\beta_k})}\eqn{x_{jk}|10^{\beta_k})^\nu
\kappa_{\nu}(2\sqrt{\nu}|x_{ik} - }{R_{ij} = \prod
\frac{1}{\Gamma(\nu)2^{\nu-1}}(2\sqrt{\nu}|x_{ik} - x_{jk}|10^{\beta_k})^\nu
\kappa_{\nu}(2\sqrt{\nu}|x_{ik} -
x_{jk}|10^{\beta_k})}\eqn{x_{jk}|10^{\beta_k})}{R_{ij} = \prod
\frac{1}{\Gamma(\nu)2^{\nu-1}}(2\sqrt{\nu}|x_{ik} - x_{jk}|10^{\beta_k})^\nu
\kappa_{\nu}(2\sqrt{\nu}|x_{ik} - x_{jk}|10^{\beta_k})},

where \eqn{\kappa_{\nu}}{\kappa_{\nu}} is the modified Bessel function of
order \eqn{\nu}{\nu}.\cr
}
\note{
Both Matern and power exponential correlation functions use the new
\eqn{\beta} parametrization of hyper-parameters given by \eqn{\theta_k =
10^{\beta_k}} for easier likelihood optimization.  That is, \code{beta} is a
log scale parameter (see MacDonald et al. (2015)).
}
\examples{
## 1D Example - 1
n = 5
d = 1
set.seed(3)
library(lhs)
x = maximinLHS(n,d)
beta =  rnorm(1)
corr_matrix(x,beta)

## 1D Example - 2
beta = rnorm(1)
corr_matrix(x,beta,corr = list(type = "matern"))

## 2D example - 1
n = 10
d = 2
set.seed(2)
library(lhs)
x = maximinLHS(n,d) 
beta = rnorm(2)
corr_matrix(x, beta,
    corr = list(type = "exponential", power = 2))

## 2D example - 2
beta = rnorm(2)
R = corr_matrix(x,beta,corr = list(type = "matern", nu = 5/2))
print(R)

}
\references{
MacDonald, K.B., Ranjan, P. and Chipman, H. (2015). 
GPfit: An R Package for Fitting a Gaussian Process Model to 
Deterministic Simulator Outputs.  
Journal of Statistical Software, 64(12), 1-23.
\url{http://www.jstatsoft.org/v64/i12/} \cr

Ranjan, P., Haynes, R., and Karsten, R. (2011). 
A Computationally Stable
Approach to Gaussian Process Interpolation of 
Deterministic Computer Simulation Data, 
Technometrics, 53(4), 366 - 378. \cr

Santner, T.J., Williams, B., and Notz, W. (2003), 
The design and analysis of computer experiments, 
Springer Verlag, New York.
}
\author{
Blake MacDonald, Hugh Chipman, Pritam Ranjan
}
\keyword{Correlation}
\keyword{Exponential}
\keyword{Gaussian}
\keyword{Matern}
\keyword{Power}
